Move InputCleaner helpers to their own file.
